Remove history changes from functions as prepration.
The fact is that the history should be manipulated in a separate function,
which is called independently from the GUI manipulating functions. The problem
with the current approach was that the waiting required to make the page history
changes work would require sacrificing other site performance to allow awaiting
some API calls – this will be solved by implementing the routing functions
properly as separate functions.