Fix logout and logout tests.
The former logout function left things hanging so they couldn't be
garbage collected properly – this has now been fixed. Also fixed
the scoped token invalidation to actually invalidate the scoped
token and made purging the session more thorough. The tests had
to be changed a bit to enable changing the logout function (purging
the session rendered it unreachable, so a separate reference was
required to actually check if invalidation was corretly called.)