Fix test tearDown method to check proper server shutdown.
There was a problem with the test class considering the server
shutdown upon test teardown, which meant that the server didn't
always have time to exit properly. This lead to the previous test
still hogging the port while the next one was trying to use it,
leading to an incorrect test failure. This could be deduced from
the fact that the next test was already fine to run, which would
mean that the server needed just a bit more time for the shutdown.