dev: persist sessions, shell command + vue devServer proxy.
This is hopefully helpful to others.
Persisting sessions help so that when the python server stops and starts again, the active sessions are reloaded, no need to sign in every time.
Shell command adds the 'app' to the context of a python or iPython shell. Automating setting up the env to run certain commands.
Vue devServer uses the Webpack devServer to proxy requests to the locally running backend, so we can use 'npm run serve' to see our changes immediately, without having to run a production build after every change to the frontend.