Add IndexedDB in-browser caching.
With this change, we improve performance (especially for large projects) by reducing the number of requests, as data is cached locally.

The perceptive experience of browsing is also improved, as there are shorter loading times. When navigating the relevant data is still updated in the background, so the number of requests didn't reduce so dramatically. The browsing efficiency did, though, as data is updated in the background and might be unnoticeable.

SharedObjects are not cached in the DB, and instead are loaded when needed.

Firefox's private mode doesn't offer IndexedDB, and so it's now unsupported.