Search objects and containers in a project.
From the browser view, in the container list,
typing the search box will search items that are in
the browser's IndexedDB.

Adding words to the search will narrow down results.
only digits and letters are valid search queries.
Special characters like _- or space are used to create
query words.

Results are returned within dozens of milliseconds,
even for hundreds of thousands of objects in the database.